+++ src/main/java/web/webspringapp/repository/NovelRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 import web.webspringapp.model.Author;
 import web.webspringapp.model.Novel;
-public interface NovelRepository  extends JpaRepository<Novel, Integer>{
+public interface NovelRepository  extends JpaRepository<Novel, Long>{
 }
